Replaced chained environments with a straight table for each env.
Previously each env was a table with the entry at 0 a pointer
to the next scope level out. Annotations labeled each var with
the number of scopes out that was needed to look for it. Now each
time a new scope is made, we simply calculate which variables
appear free in the lambda and copy just those values to the new
env. I added annotation on each lambda about which variables
are free, but I think that isn't even necessary because it can
calculated purely at that node.