Change annotation to indicate if a variable is defined in the script.
Previously we looked up the variable in the environment and if
it wasn't present then looked it up in the globals. Now we determine
at compile time if it is defined. If it is defined we look it up in
the env and otherwise assume it is in the globals and look it up
there.

The annotations no longer record the free variables in a lambda
because those can easily be determined at the lambda node itself
by using the freeVarsEx function.