add initial serializing code for quoted expressions
a quoted symbol will be a table with 'quote' set to true, 'list'
set to false and the value in entry 0. A list of quotes will have
'list' set to true, the car in entry 0 and the cdr in entry 1.
The serialization functions follow the linked list, printing the
values and concatenating them.

TODO: add serialization for all the literal value types.