remove local/global annotation for eval
The local/global annotation will be difficult with eval because
the evaled string is being compiled with the environment unknown,
e.g. if we have (quote y) and compile "y" it will annotate y as
a global variable, but it may be defined ealier in the file. So
we remove the annotation for now.